requirements.txt
streamlit==1.44.1
sentence-transformers==2.2.2
spacy==3.7.2
python-dotenv==1.1.0
PyMuPDF==1.22.0
numpy==2.2.5
scikit-learn
google-generativeai==0.8.5
protobuf
huggingface_hub==0.10.1